% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsFuncs.R
\name{fars_read}
\alias{fars_read}
\title{Title fars_read}
\source{
extdate/accident_year.csv.bz2
}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a csv file that may or may not exist}
}
\value{
a dplyr tbl_df, a wrapper around the data that came from the csv file.
    If the file does not exist, an error message will be printed.
}
\description{
This function accepts a csv file and returns a tbl_df wrapper around the data
An error message will be printed if the file does not exist
}
\examples{
\dontrun{x <- fars_read('myFile.csv')}
\dontrun{system.file("extdata", "accident_year.csv.bz2", package = "packFars"). }

}
