% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsFuncs.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Title fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{one or more years as an atomic value or a list}
}
\value{
Creates one or more datasets based on year number.  Returns NULL if there is an error
}
\description{
This function accepts one or more years as a list, calls the function make_filename()
with each of the years, and then populates those files with data associated with that
specific year from the main data set
}
\details{
An error will be thrown and the function will be halted if the year is invalid

Uses make_filename(year)
     fars_read(file)
}
\examples{
\dontrun{
     fars_read_years(1999)
     fars_read_years(as.list(1999, 2000, 2001))
     fars_read_years(1999:2016)
}
}
