% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsFuncs.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Title fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{One or more years, no error checking}
}
\value{
A wide data frame of counts by month and year,
}
\description{
Accepts a list of one or more years and passes that list to fars_read_years().
Receives from that function a data set of data in the month and year column
Uses various dplyr functions count the number of observations by month for a year.
}
\details{
Uses fars_read_years(years)
}
\examples{
\dontrun{
     fars_summarize_years(1999)
     fars_summarize_years(as.list(1999, 2000, 2001))
     fars_summarize_years(1999:2016)
}
}
