% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsFuncs.R
\name{make_filename}
\alias{make_filename}
\title{Title make_filename}
\source{
extdate/accident_year.csv.bz2
}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A year as a string, with no restrictions on format}
}
\value{
a filename in the format 'accident_year.csv.bz2'. As a side effect, the filename is printed by the function.
}
\description{
This helper function allows the user to create accident data files by year.
bz2 files are compressed files, usually on UNIX, for machines that don't support
the tar/tarball format
}
\examples{
newFileName <- make_filename('2017')
\dontrun{system.file("extdata", "accident_year.csv.bz2", package = "packFars")}
}
